requirements.txt
requests 
json 
pillow
futures
selenium
insightface==0.7.3